refactor : (domain 모듈) 도메인에 맞게 Fixture 생성 클래스 분리
